+++ packages/backend/src/modules/finality/analyzers/opStack/blobToData.ts
@@ -1,6 +1,6 @@
 import { Blob } from '@l2beat/shared'
 import { assert } from '@l2beat/shared-pure'
-import { byteArrFromHexStr } from './utils'
+import { byteArrFromHexStr } from '../../utils/byteArrFromHexStr'
 
 const BLOB_SIZE = 4096 * 32
 const MAX_BLOB_DATA_SIZE = (4 * 31 + 3) * 1024 - 4
