+++ packages/backend/src/modules/finality/analyzers/polygon-zkevm/PolygonZkevmT2IAnalyzer.ts
@@ -4,7 +4,7 @@ import { z } from 'zod'
 
 import { Database } from '@l2beat/database'
 import { RpcClient } from '@l2beat/shared'
-import { byteArrFromHexStr } from '../opStack/utils'
+import { byteArrFromHexStr } from '../../utils/byteArrFromHexStr'
 import { BaseAnalyzer } from '../types/BaseAnalyzer'
 import type { L2Block, Transaction } from '../types/BaseAnalyzer'
 import { decodeBatch } from './batch'
